+++ src/main/java/net/silentchaos512/gear/client/event/TooltipHandler.java
@@ -7,14 +7,13 @@ import net.minecraft.util.text.TextFormatting;
 import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModList;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.parts.IGearPart;
 import net.silentchaos512.gear.api.parts.IPartMaterial;
 import net.silentchaos512.gear.api.parts.MaterialGrade;
 import net.silentchaos512.gear.api.parts.PartType;
-import net.silentchaos512.gear.api.stats.ItemStats;
 import net.silentchaos512.gear.api.stats.ItemStat;
+import net.silentchaos512.gear.api.stats.ItemStats;
 import net.silentchaos512.gear.api.stats.StatInstance;
 import net.silentchaos512.gear.api.traits.ITrait;
 import net.silentchaos512.gear.client.KeyTracker;
@@ -83,7 +82,7 @@ public final class TooltipHandler {
         for (ITrait trait : traits.keySet()) {
             if (traitIndex < 0 || traitIndex == i) {
                 final int level = traits.get(trait);
-                event.getToolTip().add(trait.getDisplayName(level).applyTextStyle(TextFormatting.ITALIC));
+                trait.addInformation(level, event.getToolTip());
             }
             ++i;
         }